RobinBot/requirements.txt
ccxt>=4.3.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
